完成了createGroup和modifyGroup

+++ src/main/java/com/example/projecthelper/security/CustomUserDetails.java
@@ -17,7 +17,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 @Getter
 public class CustomUserDetails implements UserDetails {
 
-    private final Integer userId;
+    private final Long userId;
     private final String password;
     private final int identity;
